Create ceil_floor.py
Given a sorted array of numbers and given a value num, Find the ceiling of num which is the smallest number greater than or equal to num and floor of the num which is greatest number smaller than or equal to num. Print the ceil followed by floor separated by ‘,’. If there is only ceil or only floor or if the ceil and floor are the same, print the single value identified.